Get Health percentage

+++ coreDLL/injectable/wow/game/object/WowUnitObject.cpp
@@ -46,10 +46,10 @@ int WowUnitObject::getUnitMaxHealth() const {
 	return *reinterpret_cast<const uint32_t*>(getDescriptor() + WowGameOffsets::WowUnitObject::DescriptorOffsetMaxHealth);
 }
 
-int WowUnitObject::getUnitHealthPercentage() const {
-	int currentHealth = *reinterpret_cast<const uint32_t*>(getDescriptor() + WowGameOffsets::WowUnitObject::DescriptorOffsetHealth);
-	int maxHealth = *reinterpret_cast<const uint32_t*>(getDescriptor() + WowGameOffsets::WowUnitObject::DescriptorOffsetMaxHealth);
-	return (currentHealth / maxHealth * 100);
+float WowUnitObject::getUnitHealthPercentage() const {
+	float currentHealth = *reinterpret_cast<const uint32_t*>(getDescriptor() + WowGameOffsets::WowUnitObject::DescriptorOffsetHealth);
+	float maxHealth = *reinterpret_cast<const uint32_t*>(getDescriptor() + WowGameOffsets::WowUnitObject::DescriptorOffsetMaxHealth);
+	return (currentHealth / maxHealth * (float)100);
 }
 
 int WowUnitObject::getUnitEnergy() const {
